django/requirements.txt
asgiref==3.6.0
Django==4.1.2
djangorestframework==3.14.0
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
numpy==1.24.1
